+++ src/AntoCommon.h
@@ -1,0 +1,6 @@
+#ifndef ANTO_COMMON_H
+#define ANTO_COMMON_H
+
+#define ANTO_VER    "v0.5.2"
+
+#endif /* ifndef ANTO_COMMON_H */